move runner types to a separate file
