+++ runner/cmd.go
@@ -5,42 +5,12 @@ import (
 	"os"
 	"strings"
 	"syscall"
-	"time"
 
 	"os/exec"
 
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
-
-// Exit constants used to know how handle the message.
-// The consumer runnig is the responsible to understand this status and handle them properly.
-const (
-	ExitTimeout     = -1
-	ExitACK         = 0
-	ExitFailed      = 1
-	ExitNACK        = 3
-	ExitNACKRequeue = 4
-	ExitRetry       = 5
-)
-
-// Runnable represent an runnable used by consumers to handle messages.
-type Runnable interface {
-	Process(context.Context, []byte) int
-}
-
-type Options struct {
-	Path string   `mapstructure:"path"`
-	Args []string `mapstructure:"args"`
-}
-
-// Config is an composition of all options and configurations used by this runnables.
-type Config struct {
-	Type         string        `mapstructure:"type"`
-	IgnoreOutput bool          `mapstructure:"ignore-output"`
-	Options      Options       `mapstructure:"options"`
-	Timeout      time.Duration `mapstructure:"timeout"`
-}
 
 type command struct {
 	cmd          string
@@ -80,18 +50,6 @@ func (c *command) Process(ctx context.Context, b []byte) int {
 		c.l.Info("message processed with output", zap.ByteString("output", output))
 	}
 	return ExitACK
-}
-
-// New create and return a Runnable based on the config type. if the type didn't exist an error is returned.
-func New(log *zap.Logger, c Config) (Runnable, error) {
-	switch c.Type {
-	case "command":
-		return newCommand(log, c)
-	}
-	return nil, errors.Errorf(
-		"Invalid Runner type (\"%s\") expecting (%s)",
-		c.Type,
-		strings.Join([]string{"command"}, ", "))
 }
 
 func newCommand(log *zap.Logger, c Config) (*command, error) {
